+++ src/types/rx-storage.interface.d.ts
@@ -117,20 +117,6 @@ export type FilledMangoQuery<RxDocType> = Override<
  */
 export type RxStorageStatics = Readonly<{
     /**
-     * Returns a hash of the given value.
-     * Used to check equalness of attachments data and other stuff.
-     * Pouchdb uses md5 but we can use whatever we want as long as each
-     * storage class returns the same hash each time.
-     */
-    hash(data: Buffer | Blob | string): Promise<string>;
-
-    /**
-     * Key of the used hash algorithm.
-     * Like 'md5' or 'sha1'.
-     */
-    hashKey: string;
-
-    /**
      * PouchDB and others have some bugs
      * and behaviors that must be worked arround
      * before querying the db.